Refactor the notion of advancing turns and players
Instead of using a turn_over flag in the player list and then
advancing the player accordingly, there is now an advanceTurn()
and advancePlayer() method. This better matches how the game works
as the turn can advance without the player.

This also fixes a bug where the turn was not advancing after the
initial card was played.